fill and remove working correctly for tank

+++ src/main/java/Store/IStoreMedium.java
@@ -5,5 +5,5 @@ import Tank.TankSubject;
 public interface IStoreMedium {
     void fill(Enum input, Integer quantity);
     Enum[] remove(Integer quantity);
-    Integer getRelativeFillState();
+    Double getRelativeFillState();
 }
